fix: id

+++ src/main/kotlin/io/github/gnuf0rce/github/model/AutoLinksMapper.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021-2022 dsstudio Technologies and contributors.
+ * Copyright 2021-2024 dsstudio Technologies and contributors.
  *
  *  此源代码的使用受 GNU AFFERO GENERAL PUBLIC LICENSE version 3 许可证的约束, 可以在以下链接找到该许可证.
  *  Use of this source code is governed by the GNU AGPLv3 license that can be found through the following link.
@@ -26,7 +26,7 @@ public open class AutoLinksMapper(parent: Url) :
 
     public open suspend fun new(link: AutoLink): AutoLink = post(context = link)
 
-    public open suspend fun get(id: Int): AutoLink = get(path = "$id")
+    public open suspend fun get(id: Long): AutoLink = get(path = "$id")
 
-    public open suspend fun delete(id: Int): Unit = delete(path = "$id")
+    public open suspend fun delete(id: Long): Unit = delete(path = "$id")
 }
