+++ src/main/kotlin/io/github/gnuf0rce/github/model/ReleasesMapper.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021-2022 dsstudio Technologies and contributors.
+ * Copyright 2021-2024 dsstudio Technologies and contributors.
  *
  *  此源代码的使用受 GNU AFFERO GENERAL PUBLIC LICENSE version 3 许可证的约束, 可以在以下链接找到该许可证.
  *  Use of this source code is governed by the GNU AGPLv3 license that can be found through the following link.
@@ -42,7 +42,7 @@ public open class ReleasesMapper(parent: Url) :
     /**
      * [get-a-release](https://docs.github.com/en/rest/releases/releases#get-a-release)
      */
-    public open suspend fun get(id: Int): Release = get(path = "$id")
+    public open suspend fun get(id: Long): Release = get(path = "$id")
 
     // endregion
 }
